rename folder datas for inital data loader
